refactor: Use root cli.js
