+++ packages/create-project/cli.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+const { run } = require('./lib/index.js');
+run();
