--- packages/create-project/lib/index.js
@@ -1,5 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const cli_1 = require("./cli");
-(0, cli_1.run)();
-//# sourceMappingURL=index.js.map
