+++ packages/create-project/package.json
@@ -2,8 +2,8 @@
   "name": "@temporalio/create",
   "version": "0.5.0",
   "description": "Create a Temporal project from template",
-  "main": "lib/index.js",
-  "bin": "lib/index.js",
+  "main": "cli.js",
+  "bin": "cli.js",
   "scripts": {
     "build": "tsc --build",
     "build.watch": "tsc --build --watch"
